+++ ReactCommon/fabric/components/activityindicator/BUCK
@@ -15,7 +15,7 @@ rn_xplat_cxx_library(
         [
             ("", "*.h"),
         ],
-        prefix = "fabric/activityindicator",
+        prefix = "fabric/components/activityindicator",
     ),
     compiler_flags = [
         "-fexceptions",
