+++ ReactCommon/fabric/components/activityindicator/primitives.h
